BABEL: Padded strings return the wrong result in LIKE clauses
When you do string comparisons by using LIKE, all characters in the pattern string are significant. Significant characters include any leading or trailing spaces.

If a comparison in a query is to return all rows with a string LIKE 'abc ' (abc followed by a single space), a row in which the value of that column is abc (abc without a space) isn't returned. However, trailing blanks, in the expression to which the pattern is matched, are ignored. If a comparison in a query is to return all rows with the string LIKE 'abc' (abc without a space), all rows that start with abc and have zero or more trailing blanks are returned.

Reference:https://github.com/MicrosoftDocs/sql-docs/blob/live/docs/t-sql/language-elements/like-transact-sql.md

Task: BABEL-2716
Author: Jungkook Lee jungkook@amazon.com
Signed-off-by: Kushaal Shroff kushaal@amazon.com
(cherry picked from commit ac995c69ecb51d7b6c7d403971e242204af9dd15)