BABEL: Computed column ISNULL(CONVERT()) crashes PG server with segfault
In PG, the typmod of a CoalesceExpr is -1 if all the arguments of the
expression are unable to agree on type and typmod. This is not the case
for T-SQL ISNULL() (which is implemented using PG COALESCE() in
Babelfish and so uses CoalesceExpr internally). In case of T-SQL
ISNULL(), the type/typmod is determined by the first not null argument.

This commit introduces a hook namely "coalesce_typmod_hook" which is
called if we encounter a T-SQL ISNULL() function. This hook returns the
typmod of the first argument in a CoalesceExpr. If the coalesce type is
numeric, then instead of sending typmod of the first argument (which is
not entirely correct since we may not be able to fit other arguments in
that scale/precision), we use the "resolve_numeric_typmod_from_exp"
method to figure out the scale/precision for us.

Also added a boolean member variable "tsql_is_null" in the
CoalesceExpr struct which is true if we encounter a T-SQL
ISNULL(); false otherwise.

Authored-by: Sharu Goel goelshar@amazon.com
Signed-off-by: Kuntal Ghosh <kuntalgh@amazon.com>
(cherry picked from commit 27917d5093d90c8b4b6cfe0b80306b3b6d9beb9a)