BABEL: Support duplicate trigger name in different databases (#50)
Previously, duplicate trigger name was not allowed in different databases.
But in tsql, different databases can have the same trigger name. Here, we
check if a trigger with the same name is in current database. If not, then the
trigger can be created. And during drop trigger, we modify the pre check
trigger in parsing stage to not throw error, and find the correct trigger to
drop with trigger and database name.

Task:BABEL-3117
Signed-off-by: Zhenye Li <zhenye@amazon.com>
(cherry picked from commit ab122807cfde7a4b33f2ea4fa0017075efbd9487)