[#19138] Docdb/AUH: Defining Macros and wait states for AUH
Summary:
Implements the basic infrastructure for keeping track of the wait states.

WaitStateInfo is the class that stores the per-rpc wait-states.
Each incoming RPC owns one such object, and it is updated at various points
in the code using a framework that is similar to util::Trace.

We define various macros to adopt wait-state info objects -- to update the thread-local wait state
to be used in other parts of the code.

WaitStateInfo largely consists of

Upgrade/Rollback safety: Safe to upgrade/downgrade. Protos only affect the new functionality -- AUH, and do not interfere with existing functionality. AUH will be unavailable if downgraded.

Test Plan:
Jenkins: compile only
ybd --cxx-test wait_state-test

Reviewers: arybochkin, jason

Subscribers: jason, hsunder, asaha, hbhanawat, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D28647